requirements.txt
streamlit>=1.30
pytest>=7.0
langchain
langchain-community
langchain-google-genai
faiss-cpu
gpt4all
python-dotenv
scikit-learn